Backend/fakenews-app.com/requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2

djangorestframework-simplejwt==4.6.0
asgiref==3.3.1
Markdown==3.3.3
PyJWT==2.0.0

django-admin-tools==0.9.0
django-environ==0.4.5
django-multiselectfield==0.1.12
django-nested-admin==3.3.3

mysqlclient==1.4.6

numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2

python-dateutil==2.8.1
pytz==2019.3

six==1.14.0
sqlparse==0.3.1

PyMySQL==0.9.3

django-cors-headers==3.6.0
